requirements.txt
bbrl==0.3
bbrl_algos==0.0.2
bbrl_gymnasium==0.3.4
gymnasium==0.29.1
hydra-core==1.3.2
matplotlib==3.8.0
numpy==1.26.4
omegaconf==2.3.0
opencv_python==4.9.0.80
pandas==2.2.2
pygame==2.5.2
torch==2.3.0
tqdm==4.66.4